reorganize runner package
