+++ pkg/runner/program_test.go
@@ -1,4 +1,4 @@
-package zero
+package runner
 
 import (
 	"testing"
@@ -50,7 +50,7 @@ func TestLoadCairoZeroProgram(t *testing.T) {
 	program, err := LoadCairoZeroProgram(cairoZeroJson)
 	require.NoError(t, err)
 
-	require.Equal(t, &Program{
+	require.Equal(t, &ZeroProgram{
 		Bytecode: []*fp.Element{
 			stringToFelt("0x01"),
 			stringToFelt("0x02"),
