+++ pkg/runner/runner_benchmark_test.go
@@ -1,4 +1,4 @@
-package zero
+package runner
 
 import (
 	"math"
