--- pkg/runners/cairo/cairo.go
@@ -1,1 +1,0 @@
-package cairo
