--- pkg/runners/cairo/cairo_test.go
@@ -1,1 +1,0 @@
-package cairo
