--- pkg/runners/cairo/program.go
@@ -1,1 +1,0 @@
-package cairo
