--- pkg/runners/cairo/program_test.go
@@ -1,1 +1,0 @@
-package cairo
